9238354db Save adyen auth id to backend_processor_txn_id
80af727cd Localisation updates from https://translatewiki.net.
4d8ce7cd4 RecurUpgrade: Use canonical cast function name
d91e89911 FraudFilter Update: Field references now work with wildcards
92f061c79 unset processor_contact_id to avoid it map to buyer id

+++ special/RecurUpgrade.php
@@ -153,7 +153,7 @@ class RecurUpgrade extends UnlistedSpecialPage {
 			? $params['upgrade_amount_other']
 			: $params['upgrade_amount'];
 
-		$amount = $donorData['amount'] + round( (double)$upgradeAmount, 2 );
+		$amount = $donorData['amount'] + round( (float)$upgradeAmount, 2 );
 		$message = [
 				'txn_type' => 'recurring_upgrade',
 				'contribution_recur_id' => $donorData['contribution_recur_id'],
